separate main with dashboard page
